+++ remove-http.php
@@ -3,9 +3,9 @@
  * Plugin Name: Remove HTTP
  * Plugin URI: https://wordpress.org/plugins/remove-http/
  * Description: Automatically remove both HTTP and HTTPS protocols from all web links.
- * Version: 1.0.1
+ * Version: 1.0.2
  * Author: Fact Maven
- * Author URI: https://www.factmaven.com/#plugins
+ * Author URI: https://www.factmaven.com
  * License: GPLv3
  */
 
@@ -15,24 +15,34 @@ if ( ! defined( 'ABSPATH' ) ) exit;
 class Fact_Maven_Remove_HTTP {
 
     public function __construct() {
+        # Display brief update notice
+        add_action( 'in_plugin_update_message-' . plugin_basename( __FILE__ ), array( $this, 'upgrade_notice' ), 10, 2 );
         # Remove HTTP and HTTPS protocols
-        add_action( 'plugins_loaded', array( $this, 'output_buffering' ), 10, 1 );
+        add_action( 'wp_loaded', array( $this, 'output_buffering' ), 10, 1 );
+    }
+
+    public function upgrade_notice( $current, $new ) {
+        // var_dump( $new ); // Debugging
+        # If the `upgrade_notice` exists in the readme, display the info
+        if ( isset( $new->upgrade_notice ) && strlen( trim( $new->upgrade_notice ) ) > 0 ) {
+            echo '<br>' . '<strong>Upgrade Notice</strong>: ' . $new->upgrade_notice;
+        }
+        # Otherwise, display a static notice
+        else {
+            echo '<br>' . '<strong>Upgrade Notice</strong>: <span class="description">Fixed issue with plugin conflicts such as Visual Composer and Revolution slider.</span>';
+        }
     }
 
     public function output_buffering() {
-        global $pagenow;
-        # If the page is not 'General Settings', proceed
-        if ( $pagenow != 'options-general.php' ) {
-            # Enable output buffering
-            ob_start( array( $this, 'remove_protocols' ) );
-        }
+        # Enable output buffering
+        ob_start( array( $this, 'remove_protocols' ) );
     }
 
     public function remove_protocols( $buffer ) {
         $content_type = NULL;
         # Check for 'Content-Type' headers only
         foreach ( headers_list() as $header ) {
-            if (strpos( strtolower( $header ), 'content-type:' ) === 0 ) {
+            if ( strpos( strtolower( $header ), 'content-type:' ) === 0 ) {
                 $pieces = explode( ':', strtolower( $header ) );
                 $content_type = trim( $pieces[1] );
                 break;
@@ -41,11 +51,7 @@ class Fact_Maven_Remove_HTTP {
         # If the content-type is 'NULL' or 'text/html', apply rewrite
         if ( is_null( $content_type ) || substr( $content_type, 0, 9 ) === 'text/html' ) {
             # Find and remove both 'http:' and 'https:' protocols
-            $replace = preg_replace( '~=\s*["\']\s*https?:(.*?)["\']~i', '="$1"', $buffer );
-            # If there is a regex error, skip overwriting content
-            if ( $replace ) {
-                $buffer = $replace;
-            }
+            $buffer = preg_replace( "/(<(script|link|base|img|form|a|meta|iframe|svg|html)([^>]*)(href|src|action|content|xmlns)=[\"'])https?:\\/\\//i", "$1//", $buffer );
         }
         # Return protocol relative links
         return $buffer;