+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-## Changelog 
-### 1.0.1 (10/10/16)
+## Changelog
+### 1.0.2 [2016-10-24]
+* Function will run once WordPress, all plugins, and the theme are fully loaded
+* Reverted back to original regex from v1.0.0
+* Upgrade notice shown when update is available
+
+**Fixes**:
+* Conflict with [Visual Composer](https://vc.wpbakery.com) plugin
+* YouTube videos in [Revolution Slider](https://revolution.themepunch.com) plugin stopped playing
+
+### 1.0.1 [2016-10-10]
 * Improve regex to find all protocols
 * Comments added for code explanation
 
-### 1.0.0 (09/05/16)
+### 1.0.0 [2016-09-05]
 * Initial release, huzzah!